Extracting WpApiClient creation into a provider
